refactor: refactor: decompose lib/plugins/google.ts (822 lines) — split Drive/Docs
